build: migrate FastFetch project
