--- GVFS/FastFetch/package.nuspec
@@ -1,19 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>FastFetch</id>
-    <version>0.1.0</version>
-    <authors>Microsoft</authors>
-    <projectUrl>https://github.com/Microsoft/GVFS</projectUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>
-    An internal tool for quickly pulling large git repos from a GVFS-enabled git server.
-    This application will be deprecated once this same functionality is added to the gvfs prefetch verb.
-    </description>
-  </metadata>
-  <files>
-    <file src="..\..\..\BuildOutput\FastFetch\bin\x64\Release\FastFetch.exe" target="tools"/>
-    <file src="..\..\..\BuildOutput\FastFetch\bin\x64\Release\*.dll" target="tools"/>
-    <file src="..\..\..\BuildOutput\FastFetch\bin\x64\Release\*.pdb" target="tools"/>
-  </files>
-</package>
